feat(todos): implement user authentication for CRUD operations

+++ FastAPI_Complete_course/FastAPI/TodoApp/routers/todos.py
@@ -5,6 +5,7 @@ from fastapi import APIRouter, Depends, HTTPException, Path
 from starlette import status
 from models import Todos
 from database import SessionLocal
+from .auth import get_current_user
 
 router = APIRouter()
 
@@ -18,6 +19,7 @@ def get_db():
         
 #La función depende de que se abra la base de datos y se cierre después de la llamada
 db_dependency = Annotated[Session, Depends(get_db)]
+user_dependency = Annotated[dict, Depends(get_current_user)]
 
 class TodoRequest(BaseModel):
     title: str = Field(min_length = 3)
@@ -26,38 +28,64 @@ class TodoRequest(BaseModel):
     completed: bool
         
 @router.get("/")
-async def read_all(db: db_dependency):
-    return db.query(Todos).all()
+async def read_all(user: user_dependency, db: db_dependency, status_code=status.HTTP_200_OK):
+    if user is None:
+        raise HTTPException(status_code=401, detail="User not authenticated")
+    
+    return db.query(Todos).filter(Todos.owner_id == user.get("id")).all()
+
 
 @router.get("/todo/{todo_id}", status_code=status.HTTP_200_OK)
-async def read_todo( db: db_dependency,todo_id: int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def read_todo(user: user_dependency, db: db_dependency,todo_id: int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=401, detail="User not authenticated")
+    
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get("id")).first()
+    
     if todo_model is None:
         raise HTTPException(status_code=404, detail="Todo not found")
+    
     return todo_model
 
+
 @router.post("/todo", status_code=status.HTTP_201_CREATED)
-async def create_todo(db: db_dependency, todo: TodoRequest):
-    todo_model = Todos(**todo.model_dump())
+async def create_todo(user: user_dependency, db: db_dependency, todo: TodoRequest):
+    
+    if user is None:
+        raise HTTPException(status_code=401, detail="User not authenticated")
+    
+    todo_model = Todos(**todo.model_dump(), owner_id=user.get("id"))
     db.add(todo_model)
     db.commit()
 
+
 @router.put("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def update_todo(db: db_dependency, todo_request: TodoRequest, todo_id: int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def update_todo(user: user_dependency, db: db_dependency, todo_request: TodoRequest, todo_id: int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=401, detail="User not authenticated")
+    
+    # \ Se usa para dividir la línea en varias líneas
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).\
+        filter(Todos.owner_id == user.get("id")).first()
+    
     if todo_model is None:
         raise HTTPException(status_code=404, detail="Todo not found")
+    
     for key, value in todo_request.model_dump().items():
         setattr(todo_model, key, value)
+    
     db.commit()
-    db.refresh(todo_model)
-    return todo_model
+
 
 @router.delete("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def delete_todo(db: db_dependency, todo_id: int = Path(gt=0)):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+async def delete_todo(user: user_dependency, db: db_dependency, todo_id: int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(status_code=401, detail="User not authenticated")
+    
+    todo_model = db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get("id")).first()
+    
     if todo_model is None:
         raise HTTPException(status_code=404, detail="Todo not found")
+    
     db.delete(todo_model)
     db.commit()
-    return todo_model
